requirements-lambda.txt
fastapi==0.109.0
mangum==0.17.0
python-multipart==0.0.6
cellpose==3.0.8
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
